pyproject.toml
[tool.poetry]
name = "expense-manager"
version = "0.1.0"
description = ""
authors = ["vusallyv <abdulla.vusal.3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
drf-yasg = "^1.21.8"
djangorestframework = "^3.15.2"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
